Updates as part of the SIS3302 driver project
- Database customization system
- TRChannelWaveforms.db renamed to TRChannelData.db and some macros
  renamed
- Permitted some customizations in DB files:
  - Unit and EGU of the time waveform
  - Timestamp format
  - Extra fields for perSecond record
  - ZNAM/ONAM of autoRestart record
  - Link processed after a new burst
  - Allow disabling clock configuration records
  - FTVL and DTYP of waveform records
  - Allow disabling data snapshot records
- TRGenericRequest.db: support driver internally starting action
- Allow the driver to override the time array calculation (via TRArmInfo).
- Add interface_mask and interrupt_mask to TRBaseConfig
- Allow the driver to disable updating cached NDArrays in the channels
  driver (via TRBaseConfig).
- Get the value of the name record from TRBaseDriver and allow the driver
  to override the default which is the port name.
- Replace TRBaseDriver::isDisarmed() with TRBaseDriver::isArmed()
  and change the semantic so it retuns false if arming was attempted
  but there was an error before startAcqisition was called.
- Change getNumPostSamplesSnapshot and getNumPrePostSamplesSnapshot
  to always return sane values regardless of what arm mode was used.
- Add TRBaseDriver::onDisarmed virtual function to notify the driver
  when disarming is completed.
- Add TRConfigParam::getSnapshotFast which does not have an assert.
- Expose TRConfigParam::desiredParamIndex to allow the driver to know
  when the desired value of a parameter is changed.
- TRWorkerThreadTask: Add default constructor and init function to
  allow creating an array of these.

+++ TRCoreApp/src/TRConfigParam.cpp
@@ -20,7 +20,7 @@ void TRConfigParam<ValueType, EffectiveValueType>::init (
     TRBaseDriver *driver, char const *base_name, EffectiveValueType invalid_value, bool internal)
 {
     assert(!m_initialized);
-    assert(driver->m_config_params.size() < driver->m_num_config_params);
+    assert(driver->m_config_params.size() < (size_t)driver->m_num_config_params);
 
     // Remember initialization info.
     m_initialized = true;
